+++ src/main/java/io/github/astrapi69/swing/menu/MenuFactory.java
@@ -173,7 +173,7 @@ public class MenuFactory
 	public static JMenuItem newJMenuItem(final @NonNull String text, final int mnemonic,
 		final char accelerator)
 	{
-		final JMenuItem menuItem = new JMenuItem(text, mnemonic);
+		final JMenuItem menuItem = newJMenuItem(text, mnemonic);
 		MenuExtensions.setCtrlAccelerator(menuItem, accelerator);
 		return menuItem;
 	}
@@ -194,8 +194,30 @@ public class MenuFactory
 	public static JMenuItem newJMenuItem(final @NonNull String text, final int mnemonic,
 		final char accelerator, final ActionListener actionListener)
 	{
-		final JMenuItem menuItem = new JMenuItem(text, mnemonic);
+		final JMenuItem menuItem = newJMenuItem(text, mnemonic);
 		MenuExtensions.setCtrlAccelerator(menuItem, accelerator);
+		menuItem.addActionListener(actionListener);
+		return menuItem;
+	}
+
+	/**
+	 * Factory method for create a <code>JMenuItem</code>.
+	 *
+	 * @param text
+	 *            the text of the <code>JMenuItem</code>
+	 * @param mnemonic
+	 *            the keyboard mnemonic for the <code>JMenuItem</code>
+	 * @param keyStroke
+	 *            The keystroke
+	 * @param actionListener
+	 *            The action listener for the <code>JMenuItem</code>
+	 * @return the new {@link JMenuItem}
+	 */
+	public static JMenuItem newJMenuItem(final @NonNull String text, final int mnemonic,
+		final KeyStroke keyStroke, final ActionListener actionListener)
+	{
+		final JMenuItem menuItem = newJMenuItem(text, mnemonic);
+		menuItem.setAccelerator(keyStroke);
 		menuItem.addActionListener(actionListener);
 		return menuItem;
 	}
@@ -214,7 +236,7 @@ public class MenuFactory
 	public static JMenuItem newJMenuItem(final @NonNull String text, final char mnemonic,
 		final char accelerator)
 	{
-		final JMenuItem menuItem = new JMenuItem(text, mnemonic);
+		final JMenuItem menuItem = newJMenuItem(text, mnemonic);
 		MenuExtensions.setCtrlAccelerator(menuItem, accelerator);
 		return menuItem;
 	}
@@ -235,7 +257,7 @@ public class MenuFactory
 	public static JMenuItem newJMenuItem(final @NonNull String text, final char mnemonic,
 		final char accelerator, final ActionListener actionListener)
 	{
-		final JMenuItem menuItem = new JMenuItem(text, mnemonic);
+		final JMenuItem menuItem = newJMenuItem(text, mnemonic);
 		MenuExtensions.setCtrlAccelerator(menuItem, accelerator);
 		menuItem.addActionListener(actionListener);
 		return menuItem;